+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.0.4] - 2019-10-27
+
 ### Added
 
 - `condescension` preset
@@ -39,7 +41,8 @@ and this project adheres to
 - Preset support
 - `ablism` preset
 
-[unreleased]: https://github.com/erbridge/reilly/compare/v0.0.3...HEAD
+[unreleased]: https://github.com/erbridge/reilly/compare/v0.0.4...HEAD
+[0.0.4]: https://github.com/erbridge/reilly/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/erbridge/reilly/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/erbridge/reilly/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/erbridge/reilly/releases/tag/v0.0.1